src_bkp/requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.7
django-classy-tags==0.8.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-markdown-deux==1.0.5
django-pagedown==1.0.5
django-taggit==0.23.0
django-taggit-templatetags==0.2.5
django-taggit-templatetags2==1.6.1
django-templatetag-sugar==1.0
django-toolbelt==0.0.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==2.6.11
markdown2==2.3.5
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
static3==0.7.0
whitenoise==3.3.1
